full solution working
