+++ src/main/java/logging/Logging.java
@@ -1,0 +1,26 @@
+package logging;
+
+import java.util.logging.*;
+
+public class Logging {
+    
+    private static final Level LOGLEVEL = Level.ALL;
+    private static final Logger logger = Logger.getLogger("com.ibm.mq.samples.jms");
+
+
+    public static void initialiseLogging() {
+        Logger defaultLogger = Logger.getLogger("");
+        Handler[] handlers = defaultLogger.getHandlers();
+        if (handlers != null && handlers.length > 0) {
+            defaultLogger.removeHandler(handlers[0]);
+        }
+
+        Handler consoleHandler = new ConsoleHandler();
+        consoleHandler.setLevel(LOGLEVEL);
+        logger.addHandler(consoleHandler);
+
+        logger.setLevel(LOGLEVEL);
+        logger.finest("Logging initialised");
+    }
+
+}
