+++ src/main/resources/input.xml
@@ -1,17 +1,8 @@
-<employees>
-    <employee id="111">
-        <firstName>Lokesh</firstName>
-        <lastName>Gupta</lastName>
-        <location>India</location>
-    </employee>
-    <employee id="222">
-        <firstName>Alex</firstName>
-        <lastName>Gussin</lastName>
-        <location>Russia</location>
-    </employee>
-    <employee id="333">
-        <firstName>David</firstName>
-        <lastName>Feezor</lastName>
-        <location>USA</location>
-    </employee>
-</employees>
+<?xml version="1.0" encoding="UTF-8"?>
+<note>
+    <id>33333</id>
+    <to>Tove-3333</to>
+    <from>Jani-333</from>
+    <heading>Reminder</heading>
+    <phone>333</phone>
+</note>
