+++ src/main/resources/mq.properties
@@ -4,5 +4,5 @@ QMGR=QM1
 APP_USER=app
 APP_PASSWORD=app
 READ_QUEUE=DEV.QUEUE.1
-WRITE_QUEUE=
+WRITE_QUEUE=DEV.QUEUE.2
 PAYLOAD=output.xml
